pyproject.toml
[project]
name = "bombardx"
version = "0.1.0"
description = "select_ok semantics in python"
readme = "README.md"
requires-python = ">=3.9"
dependencies = []

[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[tool.maturin]
# "extension-module" tells pyo3 we want to build an extension module (skips linking against libpython.so)
features = ["pyo3/extension-module"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.2.0",
]
